+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - A bug when the user saved any changes in Config that would break **user plugin functions**.
 - ShareIsland default torrent source was **wrong**. You should modify this yourself and set the torrent source to **ShareIsland**. _Alternatively, you can reset the entire config if you desire._
 - Dupes we're not be accurately detected for numerous trackers since **v0.8.0**.
+- Alignment of widgets on sandbox input window.
 
 ### Removed